+++ mrv2/lib/mrvCore/mrvHotkey.cpp
@@ -21,9 +21,7 @@ namespace mrv
     Hotkey kOpenDirectory(true, false, false, true, 'o');
     Hotkey kOpenImage(true, false, false, false, 'o');
     Hotkey kOpenSeparateAudio(false, false, false, false, 0);
-    Hotkey kOpenSingleImage(true, true, false, false, 'o');
-    Hotkey kOpenStereoImage(false, false, true, false, 'o');
-    Hotkey kOpenSession(false, true, false, false, 'o');
+    Hotkey kOpenSession(false, false, true, false, 'o');
 
     Hotkey kSaveImage(false, false, false, false, 0);
     Hotkey kSaveSequence(true, false, false, true, 's');
@@ -142,6 +140,8 @@ namespace mrv
     Hotkey kUndoDraw(true, false, false, false, 'z');
     Hotkey kRedoDraw(true, false, false, true, 'z');
 #endif
+
+    Hotkey kSwitchPenColor(false, false, false, false, 0);
 
     Hotkey kResetChanges(true, false, false, false, 'r');
     Hotkey kExposureMore(false, false, false, false, '.');
@@ -313,18 +313,22 @@ namespace mrv
     }
 
     HotkeyEntry hotkeys[] = {
+
         HotkeyEntry(_("Open Directory"), kOpenDirectory),
         HotkeyEntry(_("Open Movie or Sequence"), kOpenImage),
-        // HotkeyEntry(_("Open Single Image"), kOpenSingleImage),
-        HotkeyEntry(_("Open Session"), kOpenSession),
+        HotkeyEntry(_("Open Session"), kOpenSession, true),
         HotkeyEntry(_("Save Image"), kSaveImage),
+
         HotkeyEntry(_("Save Movie or Sequence"), kSaveSequence),
         HotkeyEntry(_("Save PDF Document"), kSavePDF),
         HotkeyEntry(_("Save Session"), kSaveSession, true),
         HotkeyEntry(_("Save Session As"), kSaveSessionAs),
+
         HotkeyEntry(_("Close Current"), kCloseCurrent),
         HotkeyEntry(_("Close All"), kCloseAll),
+
         HotkeyEntry(_("Quit Program"), kQuitProgram),
+
         HotkeyEntry(_("Center Image"), kCenterImage),
         HotkeyEntry(_("Fit Screen"), kFitScreen),
         HotkeyEntry(_("Resize Main Window to Fit"), kResizeMainWindow),
@@ -349,9 +353,11 @@ namespace mrv
         HotkeyEntry(_("Alpha Channel"), kAlphaChannel),
         HotkeyEntry(_("Flip X"), kFlipX),
         HotkeyEntry(_("Flip Y"), kFlipY),
+
         HotkeyEntry(_("Annotation Clear Frame"), kShapeFrameClear),
         HotkeyEntry(_("Annotation Clear All Frames"), kShapeFrameClear),
         HotkeyEntry(_("Annotation Frame Step Backwards"), kShapeFrameStepBack),
+
         HotkeyEntry(_("Frame Step Backwards"), kFrameStepBack),
         HotkeyEntry(_("Frame Step FPS Backwards"), kFrameStepFPSBack),
         HotkeyEntry(_("Annotation Frame Step Forwards"), kShapeFrameStepFwd),
@@ -363,29 +369,38 @@ namespace mrv
         HotkeyEntry(_("Play Forwards"), kPlayFwd),
         HotkeyEntry(_("Play Forwards / Change Speed"), kPlayFwdTwiceSpeed),
         HotkeyEntry(_("Stop"), kStop),
+        HotkeyEntry(_("First Frame"), kFirstFrame),
+        HotkeyEntry(_("Last Frame"), kLastFrame),
+
         HotkeyEntry(_("Loop Playback"), kPlaybackLoop),
         HotkeyEntry(_("Playback Once"), kPlaybackOnce),
         HotkeyEntry(_("Playback Ping Pong"), kPlaybackPingPong),
+
         HotkeyEntry(_("First Image Version"), kFirstVersionImage),
         HotkeyEntry(_("Previous Image Version"), kPreviousVersionImage),
         HotkeyEntry(_("Next Image Version"), kNextVersionImage),
         HotkeyEntry(_("Last Image Version"), kLastVersionImage),
+
         HotkeyEntry(_("Previous Image"), kPreviousImage),
         HotkeyEntry(_("Next Image"), kNextImage),
+
         HotkeyEntry(_("Previous Image Limited"), kPreviousImageLimited),
         HotkeyEntry(_("Next Image Limited"), kNextImageLimited),
+
         HotkeyEntry(_("Previous Channel"), kPreviousChannel),
         HotkeyEntry(_("Next Channel"), kNextChannel),
-        HotkeyEntry(_("First Frame"), kFirstFrame),
-        HotkeyEntry(_("Last Frame"), kLastFrame),
+
         HotkeyEntry(_("Cut Frame"), kEditCutFrame),
         HotkeyEntry(_("Copy Frame"), kEditCopyFrame),
         HotkeyEntry(_("Paste Frame"), kEditPasteFrame),
         HotkeyEntry(_("Insert Frame"), kEditInsertFrame),
+
         HotkeyEntry(_("Slice Clip"), kEditSliceClip),
         HotkeyEntry(_("Remove Clip"), kEditRemoveClip),
+
         HotkeyEntry(_("Edit Undo"), kEditUndo),
         HotkeyEntry(_("Edit Redo"), kEditRedo),
+
         HotkeyEntry(_("Toggle Menu Bar"), kToggleMenuBar),
         HotkeyEntry(_("Toggle Top Bar"), kToggleTopBar),
         HotkeyEntry(_("Toggle Pixel Bar"), kTogglePixelBar),
@@ -406,12 +421,16 @@ namespace mrv
         HotkeyEntry(
             _("Toggle Timeline Transitions"), kToggleTimelineTransitions),
         HotkeyEntry(_("Toggle Timeline Markers"), kToggleTimelineMarkers),
+
         HotkeyEntry(_("Reset Gain/Gamma"), kResetChanges),
         HotkeyEntry(_("Exposure More"), kExposureMore),
         HotkeyEntry(_("Exposure Less"), kExposureLess),
+        HotkeyEntry(_("Gamma More"), kGammaMore),
+        HotkeyEntry(_("Gamma Less"), kGammaLess),
         HotkeyEntry(_("OCIO Input Color Space"), kOCIOInputColorSpace),
         HotkeyEntry(_("OCIO Display"), kOCIODisplay),
         HotkeyEntry(_("OCIO View"), kOCIOView),
+
         HotkeyEntry(_("Scrub Mode"), kScrubMode),
         HotkeyEntry(_("Area Selection Mode"), kAreaMode),
         HotkeyEntry(_("Draw Mode"), kDrawMode),
@@ -424,11 +443,14 @@ namespace mrv
         HotkeyEntry(_("Pen Size Less"), kPenSizeLess),
         HotkeyEntry(_("Undo Draw"), kUndoDraw),
         HotkeyEntry(_("Redo Draw"), kRedoDraw),
-        HotkeyEntry(_("Gamma More"), kGammaMore),
-        HotkeyEntry(_("Gamma Less"), kGammaLess),
+        HotkeyEntry(_("Switch Pen Color"), kSwitchPenColor),
+
         HotkeyEntry(_("Set In Point"), kSetInPoint),
         HotkeyEntry(_("Set Out Point"), kSetOutPoint),
+
         HotkeyEntry(_("Toggle Black Background"), kToggleBlackBackground),
+        HotkeyEntry(_("Hud Window"), kHudToggle),
+
         HotkeyEntry(_("Toggle One Panel Only"), kToggleOnePanelOnly),
         HotkeyEntry(_("Toggle Files Panel"), kToggleReel),
         HotkeyEntry(_("Toggle Media Info Panel"), kToggleMediaInfo),
@@ -444,11 +466,11 @@ namespace mrv
         HotkeyEntry(_("Toggle Waveform Panel"), kToggleWaveform),
         HotkeyEntry(_("Toggle Environment Map Panel"), kToggleEnvironmentMap),
         HotkeyEntry(_("Toggle Preferences Panel"), kTogglePreferences),
-        HotkeyEntry(_("Toggle Hotkeys Panel"), kToggleHotkeys),
         HotkeyEntry(_("Toggle Python Panel"), kTogglePythonConsole),
         HotkeyEntry(_("Toggle Log Panel"), kToggleLogs),
-        HotkeyEntry(_("Toggle About Panel"), kToggleAbout),
-        HotkeyEntry(_("Hud Window"), kHudToggle),
+
+        HotkeyEntry(_("Toggle Hotkeys Window"), kToggleHotkeys),
+        HotkeyEntry(_("Toggle About Window"), kToggleAbout),
         // HotkeyEntry( _("Rotate Image +90 Degrees"), kRotatePlus90),
         // HotkeyEntry( _("Rotate Image -90 Degrees"), kRotateMinus90),
         HotkeyEntry("END", kGammaLess),